requirements.txt
click>=8.1.3
PyYAML>=6.0
SQLAlchemy>=2.0.2
retry>=0.9.2
setuptools>=65.5.1
dateparser>=1.1.7
requests>=2.28.2
beautifulsoup4>=4.11.2
selenium>=4.8.0
tflunifiedapi>=0.2.1